requirements.txt
Pillow~=9.2.0
pygame~=2.5.2
snoop~=0.4.3